Build fix

+++ tasks/romanov_a_crs_product/common/include/common.hpp
@@ -2,6 +2,7 @@
 
 #include <algorithm>
 #include <cstddef>
+#include <cstdint>
 #include <exception>
 #include <random>
 #include <tuple>
@@ -15,52 +16,52 @@ namespace romanov_a_crs_product {
 constexpr double kEps = 1e-6;
 
 struct Dense {
-  size_t n;
-  size_t m;
+  uint64_t n;
+  uint64_t m;
 
   std::vector<double> data;
 
-  Dense(size_t n, size_t m) : n(n), m(m), data(n * m, 0.0) {}
-  Dense(size_t n) : n(n), m(n), data(n * n, 0.0) {}
+  Dense(uint64_t n, uint64_t m) : n(n), m(m), data(n * m, 0.0) {}
+  Dense(uint64_t n) : n(n), m(n), data(n * n, 0.0) {}
 
-  size_t getRows() const {
+  uint64_t getRows() const {
     return n;
   }
 
-  size_t getCols() const {
+  uint64_t getCols() const {
     return m;
   }
 
-  double &operator()(size_t i, size_t j) {
+  double &operator()(uint64_t i, uint64_t j) {
     return data[(i * m) + j];
   }
 
-  double operator()(size_t i, size_t j) const {
+  double operator()(uint64_t i, uint64_t j) const {
     return data[(i * m) + j];
   }
 };
 
 struct CRS {
-  size_t n;
-  size_t m;
+  uint64_t n;
+  uint64_t m;
 
   std::vector<double> value;
-  std::vector<size_t> column;
-  std::vector<size_t> row_index;
+  std::vector<uint64_t> column;
+  std::vector<uint64_t> row_index;
 
   CRS() : n(0), m(0), row_index(1, 0) {}
-  CRS(size_t n, size_t m) : n(n), m(m), row_index(n + 1, 0) {}
-  CRS(size_t n) : n(n), m(n), row_index(n + 1, 0) {}
+  CRS(uint64_t n, uint64_t m) : n(n), m(m), row_index(n + 1, 0) {}
+  CRS(uint64_t n) : n(n), m(n), row_index(n + 1, 0) {}
 
-  size_t getRows() const {
+  uint64_t getRows() const {
     return n;
   }
 
-  size_t getCols() const {
+  uint64_t getCols() const {
     return m;
   }
 
-  size_t nnz() const {
+  uint64_t nnz() const {
     return value.size();
   }
 
@@ -73,7 +74,7 @@ struct CRS {
       return false;
     }
 
-    for (size_t i = 0; i < value.size(); ++i) {
+    for (uint64_t i = 0; i < value.size(); ++i) {
       if (std::abs(value[i] - other.value[i]) > kEps) {
         return false;
       }
@@ -83,25 +84,25 @@ struct CRS {
   }
 
   void transpose() {
-    size_t nnz_val = this->nnz();
+    uint64_t nnz_val = this->nnz();
 
     std::vector<double> new_value(nnz_val);
-    std::vector<size_t> new_column(nnz_val);
-    std::vector<size_t> new_row_index(m + 1, 0);
+    std::vector<uint64_t> new_column(nnz_val);
+    std::vector<uint64_t> new_row_index(m + 1, 0);
 
-    for (size_t i = 0; i < nnz_val; ++i) {
+    for (uint64_t i = 0; i < nnz_val; ++i) {
       ++new_row_index[column[i] + 1];
     }
 
-    for (size_t i = 1; i <= m; ++i) {
+    for (uint64_t i = 1; i <= m; ++i) {
       new_row_index[i] += new_row_index[i - 1];
     }
 
-    std::vector<size_t> offset = new_row_index;
-    for (size_t row = 0; row < n; ++row) {
-      for (size_t idx = row_index[row]; idx < row_index[row + 1]; ++idx) {
-        size_t col = column[idx];
-        size_t pos = offset[col]++;
+    std::vector<uint64_t> offset = new_row_index;
+    for (uint64_t row = 0; row < n; ++row) {
+      for (uint64_t idx = row_index[row]; idx < row_index[row + 1]; ++idx) {
+        uint64_t col = column[idx];
+        uint64_t pos = offset[col]++;
 
         new_value[pos] = value[idx];
         new_column[pos] = row;
@@ -134,9 +135,9 @@ struct CRS {
     std::uniform_real_distribution<double> prob(0.0, 1.0);
     std::uniform_real_distribution<double> val(-1.0, 1.0);
 
-    for (size_t row = 0; row < n; ++row) {
+    for (uint64_t row = 0; row < n; ++row) {
       row_index[row + 1] = row_index[row];
-      for (size_t col = 0; col < m; ++col) {
+      for (uint64_t col = 0; col < m; ++col) {
         if (prob(gen) < density) {
           value.push_back(val(gen));
           column.push_back(col);
@@ -146,17 +147,17 @@ struct CRS {
     }
   }
 
-  CRS ExtractRows(size_t start, size_t end) const {
+  CRS ExtractRows(uint64_t start, uint64_t end) const {
     if (start >= end || start >= n) {
       return CRS(0, m);
     }
 
     end = std::min(end, n);
-    size_t new_n = end - start;
+    uint64_t new_n = end - start;
 
-    size_t nnz_start = row_index[start];
-    size_t nnz_end = row_index[end];
-    size_t nnz_count = nnz_end - nnz_start;
+    uint64_t nnz_start = row_index[start];
+    uint64_t nnz_end = row_index[end];
+    uint64_t nnz_count = nnz_end - nnz_start;
 
     CRS result(new_n, m);
     result.value.resize(nnz_count);
@@ -166,7 +167,7 @@ struct CRS {
     std::copy(value.begin() + nnz_start, value.begin() + nnz_end, result.value.begin());
     std::copy(column.begin() + nnz_start, column.begin() + nnz_end, result.column.begin());
 
-    for (size_t i = 0; i <= new_n; ++i) {
+    for (uint64_t i = 0; i <= new_n; ++i) {
       result.row_index[i] = row_index[start + i] - nnz_start;
     }
 
@@ -178,9 +179,9 @@ struct CRS {
       return CRS();
     }
 
-    size_t total_n = 0;
-    size_t total_nnz = 0;
-    size_t m = parts[0].m;
+    uint64_t total_n = 0;
+    uint64_t total_nnz = 0;
+    uint64_t m = parts[0].m;
 
     for (const auto &part : parts) {
       total_n += part.n;
@@ -192,14 +193,14 @@ struct CRS {
     result.column.resize(total_nnz);
     result.row_index.resize(total_n + 1);
 
-    size_t row_offset = 0;
-    size_t nnz_offset = 0;
+    uint64_t row_offset = 0;
+    uint64_t nnz_offset = 0;
 
     for (const auto &part : parts) {
       std::copy(part.value.begin(), part.value.end(), result.value.begin() + nnz_offset);
       std::copy(part.column.begin(), part.column.end(), result.column.begin() + nnz_offset);
 
-      for (size_t i = 0; i <= part.n; ++i) {
+      for (uint64_t i = 0; i <= part.n; ++i) {
         result.row_index[row_offset + i] = part.row_index[i] + nnz_offset;
       }
 
@@ -216,30 +217,30 @@ inline CRS operator*(const CRS &A, const CRS &B) {
     throw std::runtime_error("Matrix dimensions do not match for multiplication!");
   }
 
-  size_t n_rows = A.getRows();
-  size_t n_cols = B.getCols();
+  uint64_t n_rows = A.getRows();
+  uint64_t n_cols = B.getCols();
 
   CRS C(n_rows, n_cols);
   CRS BT = B.getTransposed();
 
-  std::vector<std::pair<size_t, double>> row;
+  std::vector<std::pair<uint64_t, double>> row;
 
-  for (size_t i = 0; i < n_rows; ++i) {
+  for (uint64_t i = 0; i < n_rows; ++i) {
     row.clear();
-    size_t a_begin = A.row_index[i];
-    size_t a_end = A.row_index[i + 1];
+    uint64_t a_begin = A.row_index[i];
+    uint64_t a_end = A.row_index[i + 1];
 
-    for (size_t j = 0; j < n_cols; ++j) {
-      size_t b_begin = BT.row_index[j];
-      size_t b_end = BT.row_index[j + 1];
+    for (uint64_t j = 0; j < n_cols; ++j) {
+      uint64_t b_begin = BT.row_index[j];
+      uint64_t b_end = BT.row_index[j + 1];
 
-      size_t a_pos = a_begin;
-      size_t b_pos = b_begin;
+      uint64_t a_pos = a_begin;
+      uint64_t b_pos = b_begin;
 
       double sum = 0.0;
       while (a_pos < a_end && b_pos < b_end) {
-        size_t a_col = A.column[a_pos];
-        size_t b_col = BT.column[b_pos];
+        uint64_t a_col = A.column[a_pos];
+        uint64_t b_col = BT.column[b_pos];
 
         if (a_col == b_col) {
           sum += A.value[a_pos] * BT.value[b_pos];
@@ -271,13 +272,13 @@ inline CRS operator*(const CRS &A, const CRS &B) {
 }
 
 inline CRS ToCRS(const Dense &A) {
-  size_t rows = A.getRows();
-  size_t cols = A.getCols();
+  uint64_t rows = A.getRows();
+  uint64_t cols = A.getCols();
   CRS crs(rows, cols);
 
-  for (size_t i = 0; i < rows; ++i) {
+  for (uint64_t i = 0; i < rows; ++i) {
     crs.row_index[i + 1] = crs.row_index[i];
-    for (size_t j = 0; j < cols; ++j) {
+    for (uint64_t j = 0; j < cols; ++j) {
       if (std::abs(A(i, j)) > kEps) {
         crs.value.push_back(A(i, j));
         crs.column.push_back(j);
