+++ tasks/romanov_a_crs_product/seq/src/ops_seq.cpp
@@ -1,6 +1,7 @@
 #include "romanov_a_crs_product/seq/include/ops_seq.hpp"
 
 #include <cstddef>
+#include <cstdint>
 
 #include "romanov_a_crs_product/common/include/common.hpp"
 
@@ -9,7 +10,7 @@ namespace romanov_a_crs_product {
 RomanovACRSProductSEQ::RomanovACRSProductSEQ(const InType &in) {
   SetTypeOfTask(GetStaticTypeOfTask());
   GetInput() = in;
-  GetOutput() = CRS(static_cast<size_t>(0));
+  GetOutput() = CRS(static_cast<uint64_t>(0));
 }
 
 bool RomanovACRSProductSEQ::ValidationImpl() {
